requirements.txt
apache-airflow-providers-snowflake==6.5.0
apache-airflow-providers-dbt-cloud==4.4.2